Refactored RichTextEditor component for enhanced readability and optimisation.
Renamed SocialEditor to RichTextEditor to better reflect its use case. Cleaned up the code by refactoring the large RichTextEditor component into smaller, well-defined components to improve readability and maintainability.

Furthermore, extracted common functions and properties into 'types.ts' and 'extensions.ts' for reuse across multiple files and future extensible support. Introduced 'forwardRef' and 'useImperativeHandle' to allow parent components to access the functionalities.
