+++ packages/apps/spaces/ui/form/RichTextEditor/extensions.ts
@@ -1,0 +1,27 @@
+import {
+  BlockquoteExtension,
+  BoldExtension,
+  BulletListExtension,
+  FontFamilyExtension,
+  FontSizeExtension,
+  HeadingExtension,
+  ItalicExtension,
+  NodeFormattingExtension,
+  OrderedListExtension,
+  StrikeExtension,
+  UnderlineExtension,
+} from 'remirror/extensions';
+
+export const basicEditorExtensions = () => [
+  new ItalicExtension(),
+  new BoldExtension(),
+  new StrikeExtension(),
+  new UnderlineExtension(),
+  new OrderedListExtension(),
+  new BulletListExtension(),
+  new FontSizeExtension(),
+  new FontFamilyExtension(),
+  new BlockquoteExtension(),
+  new HeadingExtension(),
+  new NodeFormattingExtension(),
+];
