+++ packages/apps/spaces/ui/form/RichTextEditor/menu/ListButtons.tsx
@@ -1,0 +1,39 @@
+import { useActive, useCommands } from '@remirror/react';
+import { Flex } from '@chakra-ui/react';
+import { IconButton } from '@ui/form/IconButton';
+import OrderedList from '../../../../components/ui/media/icons/OrderedList';
+import UnorderedList from '../../../../components/ui/media/icons/UnorderedList';
+import React from 'react';
+
+export const ListButtons = () => {
+  const { toggleOrderedList, toggleBulletList, focus } = useCommands();
+  const active = useActive();
+  return (
+    <Flex gap={2}>
+      <IconButton
+        className='customeros-remirror-button'
+        bg='transparent'
+        variant='ghost'
+        aria-label='Strikethrough'
+        onClick={() => {
+          toggleOrderedList();
+          focus();
+        }}
+        isActive={active.orderedList()}
+        icon={<OrderedList currentColor='inherit' />}
+      />
+      <IconButton
+        className='customeros-remirror-button'
+        bg='transparent'
+        variant='ghost'
+        aria-label='Underline'
+        onClick={() => {
+          toggleBulletList();
+          focus();
+        }}
+        isActive={active.bulletList()}
+        icon={<UnorderedList currentColor='inherit' />}
+      />
+    </Flex>
+  );
+};
